Add standard game data class

+++ src/main/java/fr/unice/polytech/ps5/takenoko/et2/decision/bots/MinMaxBot.java
@@ -135,7 +135,7 @@ public class MinMaxBot extends DecisionMaker
                 break;
             }
         }
-        long power = (int) Math.pow(100d, (double) n);
+        long power = (int) Math.pow(100d, n);
         Board newBoard = (Board) board.clone();
         newBoard.addTile(playedTile, playedPos, cloneBambooReserv(player.getGame().getBambooReserve()));
         int scoreReturn = 0;
